refactor: organize the structure
